requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.9.0.80
matplotlib>=3.7.1
numpy>=1.24.2
Pillow>=9.4.0
tqdm>=4.65.2

